pyproject.toml
[project]
name = "bma"
version = "0.1.0"
description = "Payments and Bill Management App"
authors = []
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-tenants (>=3.9.0,<4.0.0)",
    "razorpay (>=2.0.0,<3.0.0)",
    "django-model-utils (>=5.0.0,<6.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "weasyprint (>=66.0,<67.0)",
    "django-shell-plus (>=1.1.7,<2.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "ipython (>=9.6.0,<10.0.0)",
    "jq (>=1.10.0,<2.0.0)",
    "flask (>=3.1.2,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
